+++ src/components/v2v3/shared/DiffedSplits/DiffedSplitItem.tsx
@@ -11,10 +11,10 @@ import { SplitWithDiff } from 'utils/splits'
 import { isJuiceboxProjectSplit } from 'utils/v2v3/distributions'
 import { SplitProps } from '../SplitItem'
 import { ETHAddressBeneficiary } from '../SplitItem/EthAddressBeneficiary'
-import { LockedUntilValue } from '../SplitItem/LockedUntilValue'
 import { ReservedTokensValue } from '../SplitItem/ReservedTokensValue'
 import { SplitValue } from '../SplitItem/SplitValue'
 import { DiffedJBProjectBeneficiary } from './DiffedSplitFields/DiffedJBProjectBeneficiary'
+import { DiffedLockedUntil } from './DiffedSplitFields/DiffedLockedUntil'
 
 type DiffedSplitProps = Omit<SplitProps, 'split'> & { split: SplitWithDiff }
 
@@ -55,10 +55,9 @@ export function DiffedSplitItem({ props }: { props: DiffedSplitProps }) {
         </div>
 
         {(props.split.lockedUntil && props.split.lockedUntil > 0) || hasDiff ? (
-          <LockedUntilValue
+          <DiffedLockedUntil
             lockedUntil={props.split.lockedUntil}
             oldLockedUntil={hasDiff ? oldSplit.lockedUntil : undefined}
-            showDiff={hasDiff} // need for LockedText because lockedUntil can be undefined
           />
         ) : null}
       </div>
